Fix decoding of pid_t argument of prlimit64 syscall
* resource.c (SYS_FUNC(prlimit64)): Print pid_t syscall
argument using %d format.

+++ resource.c
@@ -142,7 +142,7 @@ SYS_FUNC(setrlimit)
 SYS_FUNC(prlimit64)
 {
 	if (entering(tcp)) {
-		tprintf("%ld, ", tcp->u_arg[0]);
+		tprintf("%d, ", (int) tcp->u_arg[0]);
 		printxval(resources, tcp->u_arg[1], "RLIMIT_???");
 		tprints(", ");
 		print_rlimit64(tcp, tcp->u_arg[2]);
